extract MonacoService from BaseEditor class
