+++ src/platform/editor/base-editor.ts
@@ -1,6 +1,7 @@
 import { AfterViewInit, ElementRef, EventEmitter, Input, OnDestroy, Output, ViewChild } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { NgxMonacoEditorConfig } from './config';
+import { MonacoService } from './monaco.service';
 
 let loadedMonaco: boolean = false;
 let loadPromise: Promise<void>;
@@ -26,46 +27,13 @@ export abstract class BaseEditor implements AfterViewInit, OnDestroy {
     return this._options;
   }
 
-  constructor(private config: NgxMonacoEditorConfig) {}
+  constructor(private monacoService: MonacoService, private config: NgxMonacoEditorConfig) {}
 
   ngAfterViewInit(): void {
-    if (loadedMonaco) {
-      // Wait until monaco editor is available
-      loadPromise.then(() => {
-        this.initMonaco(this.options);
-      });
-    } else {
-      loadedMonaco = true;
-      loadPromise = new Promise<void>((resolve: any) => {
-        const baseUrl = this.config.baseUrl || '/assets';
-        if (typeof((<any>window).monaco) === 'object') {
-          resolve();
-          return;
-        }
-        const onGotAmdLoader: any = () => {
-          // Load monaco
-          (<any>window).require.config({ paths: { 'vs': `${baseUrl}/monaco/vs` } });
-          (<any>window).require(['vs/editor/editor.main'], () => {
-            if (typeof this.config.onMonacoLoad === 'function') {
-              this.config.onMonacoLoad();
-            }
-            this.initMonaco(this.options);
-            resolve();
-          });
-        };
-
-        // Load AMD loader if necessary
-        if (!(<any>window).require) {
-          const loaderScript: HTMLScriptElement = document.createElement('script');
-          loaderScript.type = 'text/javascript';
-          loaderScript.src = `${baseUrl}/monaco/vs/loader.js`;
-          loaderScript.addEventListener('load', onGotAmdLoader);
-          document.body.appendChild(loaderScript);
-        } else {
-          onGotAmdLoader();
-        }
-      });
-    }
+    // Wait until monaco editor is available
+    this.monacoService.loadMonaco().then(() => {
+      this.initMonaco(this.options);
+    });
   }
 
   protected abstract initMonaco(options: any): void;
